+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @cat-factory/provider-s3
 
+## 0.2.36
+
+### Patch Changes
+
+- @cat-factory/kernel@0.61.1
+
+## 0.2.35
+
+### Patch Changes
+
+- Updated dependencies [15c5894]
+  - @cat-factory/kernel@0.61.0
+
 ## 0.2.34
 
 ### Patch Changes