+++ backend/packages/sandbox-fixtures/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/sandbox-fixtures
 
+## 0.7.70
+
+### Patch Changes
+
+- Updated dependencies [9bb75b0]
+  - @cat-factory/contracts@0.64.0
+
+## 0.7.69
+
+### Patch Changes
+
+- Updated dependencies [15c5894]
+  - @cat-factory/contracts@0.63.0
+
 ## 0.7.68
 
 ### Patch Changes